move tests to userauth_test.go & reduce comments
